refactor(backend): move proto rule group to proto_library_target (M2 of #1264)
Migrate the proto_library ninja rules (proto / protojava / protopython /
protodescriptors / protogo) and the protoc_import_path_option helper out of
backend.py to proto_library_target.py, registered via the rule registry
(ORDER_PROTO). The provider emits the protoc* flag vars via ctx.add_line and
the rules via ctx.emit_rule(NinjaRule(...)); config/go-plugin logic is moved
verbatim.

- proto_library_target: _generate_proto_rules provider + protoc_import_path_option
- backend.py: delete generate_proto_rules + protoc_import_path_option + proto lambda

Verified: build.ninja rule section byte-identical on blade-test (MSVC); registry
+ lex_yacc unit tests green; clean proto_basic build passes.
